test(runner): consolidate retry coverage
